pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quicksave"
version = "0.1.0"
description = "A command line tool for saving snapshots of game saves"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "boto3>=1.38.30",
    "pyyaml>=6.0.2",
]
authors = [
    {name = "Mathias Staricka"}
]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
quicksave = "quicksave.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["quicksave"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
]

